requirements.txt
asyncio
grpcio>=1.50.0
grpcio-tools>=1.50.0
pinecone-client>=2.2.0
pytest>=7.0.0
pytest-cov>=4.0.0
bandit>=1.7.0
pyyaml>=6.0
numpy>=1.24.0
aiohttp>=3.8.0